+++ Packages/com.vareiko.foundation/Runtime/Observability/FoundationObservabilityInstaller.cs
@@ -1,3 +1,5 @@
+using System.IO;
+using UnityEngine;
 using Zenject;
 
 namespace Vareiko.Foundation.Observability
@@ -23,6 +25,8 @@ namespace Vareiko.Foundation.Observability
 
             container.DeclareSignal<LogMessageEmittedSignal>();
             container.DeclareSignal<DiagnosticsSnapshotUpdatedSignal>();
+            container.DeclareSignal<DiagnosticsSnapshotExportedSignal>();
+            container.DeclareSignal<DiagnosticsSnapshotExportFailedSignal>();
             container.DeclareSignal<UnhandledExceptionCapturedSignal>();
             container.DeclareSignal<CrashReportSubmittedSignal>();
             container.DeclareSignal<CrashReportSubmissionFailedSignal>();
@@ -32,8 +36,15 @@ namespace Vareiko.Foundation.Observability
                 container.Bind<IFoundationLogSink>().To<UnityConsoleLogSink>().AsSingle().IfNotBound();
             }
 
+            container.Bind<string>()
+                .WithId("DiagnosticsExportRootPath")
+                .FromInstance(Path.Combine(Application.persistentDataPath, "foundation-diagnostics"))
+                .AsSingle()
+                .IfNotBound();
+
             container.Bind<IFoundationLogger>().To<UnityFoundationLogger>().AsSingle();
             container.BindInterfacesAndSelfTo<FoundationDiagnosticsService>().AsSingle().NonLazy();
+            container.Bind<IDiagnosticsSnapshotExportService>().To<DiagnosticsSnapshotExportService>().AsSingle();
             container.BindInterfacesAndSelfTo<GlobalExceptionHandler>().AsSingle().NonLazy();
         }
     }